backend/requirements.txt
# === Core ===
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*

# === Database & Cache ===
supabase==2.13.*
redis==5.2.*

# === HTTP Client (pour les providers) ===
httpx==0.28.*

# === Security ===
cryptography==44.*

# === Schema handling ===
fastavro==1.10.*       # Avro parsing & validation
jsonschema==4.23.*     # JSON Schema validation

# === AsyncAPI ===
pyyaml==6.0.*

# === Dev & Testing ===
pytest==8.3.*
pytest-asyncio==0.25.*
pytest-cov==6.0.*
ruff==0.9.*            # Linter + formatter
pre-commit==4.1.*

# === Utilities ===
loguru==0.7.*
python-dotenv==1.0.*
